requirements.txt
chromadb==1.0.16
Crawl4AI==0.7.3
googlesearch-python==1.3.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-ollama==0.3.6
ollama==0.5.3
pandas==2.3.1
playwright==1.54.0
streamlit==1.48.0
